Resolve header orders
